Added support for logging, fault tolerance and rate limiting

+++ src/main/java/com/student/mgmt/service/StudentService.java
@@ -3,6 +3,8 @@ package com.student.mgmt.service;
 import java.util.List;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import com.student.mgmt.entity.Student;
 import com.student.mgmt.exception.EmailAlreadyExistsException;
@@ -14,36 +16,47 @@ import com.student.mgmt.util.StudentMapper;
 @Service
 public class StudentService {
 
+    private static final Logger logger = LoggerFactory.getLogger(StudentService.class);
+
     private StudentRepository studentRepository;
 
     public StudentService(StudentRepository studentRepository) {
         this.studentRepository = studentRepository;
     }
 
-    public List<StudentDto> getAllStudents() {
+    public Optional<List<StudentDto>> getAllStudents() {
+        logger.info("Fetching all students");
         List<Student> students = studentRepository.findAll();
-        return StudentMapper.convertToDTOList(students);
+        if(students.isEmpty()) {
+            logger.warn("No students found");
+            throw new StudentNotFoundException("No students found");
+        }
+        return Optional.ofNullable(StudentMapper.convertToDTOList(students));
     }
 
-    public StudentDto getStudentById(Long id) {
+    public Optional<StudentDto> getStudentById(Long id) {
+        logger.info("Fetching student with ID: {}", id);
         Student student = studentRepository.findById(id)
                 .orElseThrow(() -> new StudentNotFoundException("Student not found"));
-        return StudentMapper.convertToDto(student);
+        return Optional.ofNullable(StudentMapper.convertToDto(student));
     }
 
-    public StudentDto getStudentByEmail(String email) {
+    public Optional<StudentDto> getStudentByEmail(String email) {
+        logger.info("Fetching student with email: {}", email);
         Student student = studentRepository.findByEmail(email)
                 .orElseThrow(() -> new StudentNotFoundException("Student not found"));
-        return StudentMapper.convertToDto(student);
+        return Optional.ofNullable(StudentMapper.convertToDto(student));
     }
     
     public StudentDto createStudent(StudentDto studentDto) {
+        logger.info("Creating new student: {}", studentDto);
         Student student = StudentMapper.convertToEntitySave(studentDto);
         student = studentRepository.save(student);
         return StudentMapper.convertToDto(student);
     }
 
-    public StudentDto updateStudent(Long id, StudentDto studentDto) {
+    public Optional<StudentDto> updateStudent(Long id, StudentDto studentDto) {
+        logger.info("Updating student with ID: {}", id);
         Student student = studentRepository.findById(id)
                 .orElseThrow(() -> new StudentNotFoundException("Student not found"));
 
@@ -57,10 +70,11 @@ public class StudentService {
         student.setGender(studentDto.getGender());
         student.setAge(studentDto.getAge());
         student.setNationality(studentDto.getNationality());       
-        return StudentMapper.convertToDto(studentRepository.save(student));
+        return Optional.ofNullable(StudentMapper.convertToDto(studentRepository.save(student)));
     }
 
-    public StudentDto updateStudentByEmail(Long id, String email) {
+    public Optional<StudentDto> updateStudentByEmail(Long id, String email) {
+        logger.info("Updating student with ID: {} and email: {}", id, email);
         Student student = studentRepository.findById(id)
                 .orElseThrow(() -> new StudentNotFoundException("Student not found"));
         
@@ -69,11 +83,16 @@ public class StudentService {
         }
 
         student.setEmail(email);
-        return StudentMapper.convertToDto(studentRepository.save(student));
+        return Optional.ofNullable(StudentMapper.convertToDto(studentRepository.save(student)));
     }
 
     public void deleteStudent(Long id) {
-        studentRepository.deleteById(id);
+        logger.info("Deleting student with ID: {}", id);
+        Student student = studentRepository.findById(id)
+                .orElseThrow(() -> new StudentNotFoundException("Student not found"));
+
+        if(student != null) {
+            studentRepository.deleteById(id);
+        }
     }
- 
 }
